requirements.txt
pyodbc>=5.3.0
python-dotenv>=1.2.1

